+++ bot/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed a real number required error
+- Fixed incorrect work of the purchase command
 
 ## 1.77 - 2023-08-06